+++ ui/src/lib/config.ts
@@ -5,6 +5,7 @@ export type UIConfig = {
   prometheus: string
   grafana: string
   wiremock: string
+  redis: string
   stompUrl: string
   stompUser: string
   stompPasscode: string
@@ -22,8 +23,9 @@ const config: UIConfig = {
   rabbitmq: `/rabbitmq`,
   prometheus: `/prometheus`,
   grafana: `/grafana`,
+  redis: `/redis`,
   // grafana: `http://${host}:3333`,
-  wiremock: `http://${host}:8080/__admin/`,
+  wiremock: `http://${host}:8080/__admin/webapp/`,
   stompUrl: `/ws`,
   stompUser: readOnlyUser,
   stompPasscode: readOnlyPasscode,